- Win: fixed compilation after object editor changes + some warnings. - TUT:   - Compilation fixes.   - Added base path setup depending on architecture to make test run at all.   - Removed flag for the autocompletion cache test that waits for feedback (feedback has long gone)   - Added fixes for all parser related tests where necessary (except for auto completion, which is to be changed radically quite soon). - Auto completion: switched to new parser services (actually, it uses the parser context from its owning editor). - Win: refactored all remaining object editor wrappers from *BE to *Wrapper. - Table editor: factored out trigger add/delete into own routines.
[TESTS]
  - Ran all parser related TUT tests.

+++ backend/wbpublic/grtsqlparser/mysql_parser_services.cpp
@@ -36,7 +36,7 @@ ParserContext::ParserContext(GrtCharacterSetsRef charsets, GrtVersionRef version
 
   // 3 character sets were added in version 5.5.3. Remove them from the list if the current version
   // is lower than that.
-  long server_version;
+  ssize_t server_version;
   if (version.is_valid())
   {
     server_version = version->majorNumber() * 10000;
@@ -56,7 +56,7 @@ ParserContext::ParserContext(GrtCharacterSetsRef charsets, GrtVersionRef version
     filtered_charsets.erase("utf32");
   }
   
-  _recognizer = new MySQLRecognizer(server_version, "", filtered_charsets);
+  _recognizer = new MySQLRecognizer((long)server_version, "", filtered_charsets);
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -79,7 +79,7 @@ void ParserContext::use_server_version(GrtVersionRef version)
 {
   _version = version;
   
-  long server_version;
+  ssize_t server_version;
   if (version.is_valid())
   {
     server_version = version->majorNumber() * 10000;
@@ -92,7 +92,7 @@ void ParserContext::use_server_version(GrtVersionRef version)
   }
   else
     server_version = 50501; // Assume some reasonable default (5.5.1).
-  _recognizer->set_server_version(server_version);
+  _recognizer->set_server_version((long)server_version);
 }
 
 //--------------------------------------------------------------------------------------------------
